requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2
django-heroku==0.3.1
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-smartfields==1.1.3
gunicorn==20.1.0
psycopg2==2.8.6
python-decouple==3.4
Pillow==8.2.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
whitenoise==5.2.0